Enemies Updates + Potion Mechanic (No Visuals Only Core Mechanic)

+++ Project Smash/Assets/_Scripts/Items/Doors/Door.cs
@@ -54,7 +54,7 @@ namespace PSmash.Items.Doors
             EnablePlayerController(false);
         }
 
-        public void TakeDamage(Transform attacker, int damage)
+        public void TakeDamage(Transform attacker, float damage)
         {
             audioSource.PlayOneShot(hitOnDoorSound);
         }
